test(slack-bridge): split fortress into package + nit cleanup (#1066)
Post-merge follow-ups from the independent review of #1064.

- Nit 1: split tests/integration/test_slack_bridge_e2e.py (1243 LOC,
  over the 500-line epic ceiling) into the slack_bridge_e2e/ package,
  mirroring the tests/teatree_core/management_commands/ convention:
  conftest.py for shared infra (FakeSlackTransport, transport fixture,
  fake config, _own_loop) + per-concern test_inbound / test_outbound /
  test_routing / test_backend_error_paths. Pure cut-paste of test
  bodies; per-file pytestmark duplicated since it is not conftest-
  inherited.
- Nit 2: module-level comment in conftest.py and one-liners above the
  per-overlay routing patch stacks explaining the deliberate config-
  resolution patching (httpx + read_pass stay the only real externals).
- Nit 3: TestxoxpVsxoxbRouting -> TestXoxpVsXoxbRouting.
- Nit 5: drop the stray module-level __all__ (zero importers).
- Nit 6: comment above the consumer-namespace messaging_from_overlay
  patch pinning down why the consumer namespace is the right target.

Behaviour-preserving: 51 collected == 51 baseline, nodeid parity
(only the intentional Nit 3 rename differs), per-module coverage on
backend_factory/slack_bot/notify unchanged, snapshots stable with
--inline-snapshot=disable.
